+++ src/aux-common/partitions/RemoteCausalRepoPartition.spec.ts
@@ -14,6 +14,8 @@ import {
     RECEIVE_EVENT,
     COMMIT,
     WATCH_COMMITS,
+    GET_BRANCH,
+    WATCH_BRANCH,
 } from '@casual-simulation/causal-trees/core2';
 import {
     remote,
@@ -24,8 +26,17 @@ import {
 } from '@casual-simulation/causal-trees';
 import flatMap from 'lodash/flatMap';
 import { waitAsync } from '../test/TestHelpers';
-import { botAdded, createBot, botUpdated, Bot, UpdatedBot } from '../bots';
-import { AuxOpType, bot, tag, value } from '../aux-format-2';
+import {
+    botAdded,
+    createBot,
+    botUpdated,
+    Bot,
+    UpdatedBot,
+    unlockSpace,
+    asyncResult,
+    asyncError,
+} from '../bots';
+import { AuxOpType, bot, tag, value, AuxCausalTree } from '../aux-format-2';
 import { RemoteCausalRepoPartitionConfig } from './AuxPartitionConfig';
 
 console.log = jest.fn();
@@ -94,8 +105,18 @@ describe('RemoteCausalRepoPartition', () => {
             sub.unsubscribe();
         });
 
-        it('should return immediate for the editStrategy', () => {
+        it('should return immediate for the realtimeStrategy if the partition is not static', () => {
             expect(partition.realtimeStrategy).toEqual('immediate');
+        });
+
+        it('should return delayed for the realtimeStrategy if the partition is static', () => {
+            setupPartition({
+                type: 'remote_causal_repo',
+                branch: 'testBranch',
+                host: 'testHost',
+                static: true,
+            });
+            expect(partition.realtimeStrategy).toEqual('delayed');
         });
 
         describe('remote events', () => {
@@ -157,6 +178,28 @@ describe('RemoteCausalRepoPartition', () => {
                     branch: 'testBranch',
                     host: 'testHost',
                     readOnly: true,
+                });
+
+                await partition.sendRemoteEvents([
+                    remote(
+                        {
+                            type: 'def',
+                        },
+                        {
+                            deviceId: 'device',
+                        }
+                    ),
+                ]);
+
+                expect(connection.sentMessages).toEqual([]);
+            });
+
+            it('should not send events when in static mode', async () => {
+                setupPartition({
+                    type: 'remote_causal_repo',
+                    branch: 'testBranch',
+                    host: 'testHost',
+                    static: true,
                 });
 
                 await partition.sendRemoteEvents([
@@ -347,8 +390,6 @@ describe('RemoteCausalRepoPartition', () => {
                 expect(removed).toEqual([]);
                 expect(updated).toEqual([]);
             });
-
-            // it('should not add the given atoms when ')
         });
 
         describe('atoms', () => {
@@ -366,6 +407,367 @@ describe('RemoteCausalRepoPartition', () => {
                 await waitAsync();
 
                 expect(connection.sentMessages.slice(1)).toEqual([]);
+            });
+
+            it('should not send new atoms to the server if in static mode', async () => {
+                setupPartition({
+                    type: 'remote_causal_repo',
+                    branch: 'testBranch',
+                    host: 'testHost',
+                    static: true,
+                });
+
+                partition.connect();
+
+                await partition.applyEvents([botAdded(createBot('bot1'))]);
+                await waitAsync();
+
+                expect(connection.sentMessages.slice(1)).toEqual([]);
+            });
+        });
+
+        describe('static mode', () => {
+            it('should send a GET_BRANCH event when in static mode', async () => {
+                setupPartition({
+                    type: 'remote_causal_repo',
+                    branch: 'testBranch',
+                    host: 'testHost',
+                    static: true,
+                });
+
+                expect(connection.sentMessages).toEqual([]);
+                partition.connect();
+
+                await waitAsync();
+
+                expect(connection.sentMessages).toEqual([
+                    {
+                        name: GET_BRANCH,
+                        data: 'testBranch',
+                    },
+                ]);
+            });
+
+            it('should not apply atoms to the causal tree', async () => {
+                setupPartition({
+                    type: 'remote_causal_repo',
+                    branch: 'testBranch',
+                    host: 'testHost',
+                    static: true,
+                });
+
+                expect(connection.sentMessages).toEqual([]);
+                partition.connect();
+
+                const ret = await partition.applyEvents([
+                    botAdded(
+                        createBot('test', {
+                            abc: 'def',
+                        })
+                    ),
+                ]);
+
+                expect(ret).toEqual([]);
+                expect(partition.state).toEqual({});
+            });
+
+            it('should load the initial state properly', async () => {
+                setupPartition({
+                    type: 'remote_causal_repo',
+                    branch: 'testBranch',
+                    host: 'testHost',
+                    static: true,
+                });
+
+                const bot1 = atom(atomId('a', 1), null, bot('bot1'));
+                const tag1 = atom(atomId('a', 2), bot1, tag('tag1'));
+                const value1 = atom(atomId('a', 3), tag1, value('abc'));
+
+                partition.connect();
+
+                addAtoms.next({
+                    branch: 'testBranch',
+                    atoms: [bot1, tag1, value1],
+                });
+
+                expect(partition.state).toEqual({
+                    bot1: createBot('bot1', {
+                        tag1: 'abc',
+                    }),
+                });
+            });
+
+            it('should transition to non static when a unlock_space event is sent', async () => {
+                setupPartition({
+                    type: 'remote_causal_repo',
+                    branch: 'testBranch',
+                    host: 'testHost',
+                    static: true,
+                });
+
+                const bot1 = atom(atomId('a', 1), null, bot('bot1'));
+                const tag1 = atom(atomId('a', 2), bot1, tag('tag1'));
+                const value1 = atom(atomId('a', 3), tag1, value('abc'));
+
+                partition.connect();
+
+                addAtoms.next({
+                    branch: 'testBranch',
+                    atoms: [bot1, tag1, value1],
+                });
+
+                await partition.applyEvents([
+                    unlockSpace(<any>'admin', '3342'),
+                    botAdded(
+                        createBot('test1', {
+                            hello: 'world',
+                        })
+                    ),
+                ]);
+
+                expect(partition.state).toEqual({
+                    bot1: createBot('bot1', {
+                        tag1: 'abc',
+                    }),
+                    test1: createBot('test1', {
+                        hello: 'world',
+                    }),
+                });
+            });
+
+            it('should transition to non read only when a unlock_space event is sent', async () => {
+                setupPartition({
+                    type: 'remote_causal_repo',
+                    branch: 'testBranch',
+                    host: 'testHost',
+                    static: true,
+                    readOnly: true,
+                });
+
+                const bot1 = atom(atomId('a', 1), null, bot('bot1'));
+                const tag1 = atom(atomId('a', 2), bot1, tag('tag1'));
+                const value1 = atom(atomId('a', 3), tag1, value('abc'));
+
+                const tree = (<any>partition)._tree as AuxCausalTree;
+                const test1 = atom(atomId(tree.site.id, 5), null, bot('test1'));
+                const helloTag = atom(
+                    atomId(tree.site.id, 6),
+                    test1,
+                    tag('hello')
+                );
+                const worldValue = atom(
+                    atomId(tree.site.id, 7),
+                    helloTag,
+                    value('world')
+                );
+
+                partition.connect();
+
+                addAtoms.next({
+                    branch: 'testBranch',
+                    atoms: [bot1, tag1, value1],
+                });
+
+                await partition.applyEvents([
+                    unlockSpace('admin', '3342'),
+                    botAdded(
+                        createBot('test1', {
+                            hello: 'world',
+                        })
+                    ),
+                ]);
+
+                expect(connection.sentMessages.slice(2)).toEqual([
+                    {
+                        name: ADD_ATOMS,
+                        data: {
+                            branch: 'testBranch',
+                            atoms: [test1, helloTag, worldValue],
+                        },
+                    },
+                ]);
+            });
+
+            it('should not transition when a unlock_space event is sent with the wrong password', async () => {
+                setupPartition({
+                    type: 'remote_causal_repo',
+                    branch: 'testBranch',
+                    host: 'testHost',
+                    static: true,
+                    readOnly: true,
+                });
+
+                const bot1 = atom(atomId('a', 1), null, bot('bot1'));
+                const tag1 = atom(atomId('a', 2), bot1, tag('tag1'));
+                const value1 = atom(atomId('a', 3), tag1, value('abc'));
+
+                partition.connect();
+
+                addAtoms.next({
+                    branch: 'testBranch',
+                    atoms: [bot1, tag1, value1],
+                });
+
+                await partition.applyEvents([
+                    unlockSpace('admin', 'wrong'),
+                    botAdded(
+                        createBot('test1', {
+                            hello: 'world',
+                        })
+                    ),
+                ]);
+
+                expect(connection.sentMessages.slice(2)).toEqual([]);
+            });
+
+            it('should not try to connect if it is not already connected', async () => {
+                setupPartition({
+                    type: 'remote_causal_repo',
+                    branch: 'testBranch',
+                    host: 'testHost',
+                    static: true,
+                    readOnly: true,
+                });
+
+                await partition.applyEvents([unlockSpace('admin', '3342')]);
+
+                expect(connection.sentMessages.length).toEqual(0);
+            });
+
+            it('should be able to unlock while waiting for the initial connection to finish', async () => {
+                setupPartition({
+                    type: 'remote_causal_repo',
+                    branch: 'testBranch',
+                    host: 'testHost',
+                    static: true,
+                    readOnly: true,
+                });
+
+                const bot1 = atom(atomId('a', 1), null, bot('bot1'));
+                const tag1 = atom(atomId('a', 2), bot1, tag('tag1'));
+                const value1 = atom(atomId('a', 3), tag1, value('abc'));
+
+                partition.connect();
+
+                await partition.applyEvents([unlockSpace('admin', '3342')]);
+
+                expect(connection.sentMessages).toEqual([
+                    {
+                        name: GET_BRANCH,
+                        data: 'testBranch',
+                    },
+                ]);
+
+                addAtoms.next({
+                    branch: 'testBranch',
+                    atoms: [bot1, tag1, value1],
+                });
+
+                expect(connection.sentMessages.slice(1)).toEqual([
+                    {
+                        name: WATCH_BRANCH,
+                        data: 'testBranch',
+                    },
+                ]);
+            });
+
+            it('should resolve the async task when unlocked', async () => {
+                setupPartition({
+                    type: 'remote_causal_repo',
+                    branch: 'testBranch',
+                    host: 'testHost',
+                    static: true,
+                    readOnly: true,
+                });
+
+                const bot1 = atom(atomId('a', 1), null, bot('bot1'));
+                const tag1 = atom(atomId('a', 2), bot1, tag('tag1'));
+                const value1 = atom(atomId('a', 3), tag1, value('abc'));
+
+                partition.connect();
+
+                addAtoms.next({
+                    branch: 'testBranch',
+                    atoms: [bot1, tag1, value1],
+                });
+
+                let events = [] as Action[];
+                partition.onEvents.subscribe(e => events.push(...e));
+                await partition.applyEvents([
+                    unlockSpace('admin', '3342', 123),
+                ]);
+
+                await waitAsync();
+
+                expect(events).toEqual([asyncResult(123, undefined)]);
+            });
+
+            it('should resolve the async task if already unlocked', async () => {
+                setupPartition({
+                    type: 'remote_causal_repo',
+                    branch: 'testBranch',
+                    host: 'testHost',
+                });
+
+                const bot1 = atom(atomId('a', 1), null, bot('bot1'));
+                const tag1 = atom(atomId('a', 2), bot1, tag('tag1'));
+                const value1 = atom(atomId('a', 3), tag1, value('abc'));
+
+                partition.connect();
+
+                addAtoms.next({
+                    branch: 'testBranch',
+                    atoms: [bot1, tag1, value1],
+                });
+
+                let events = [] as Action[];
+                partition.onEvents.subscribe(e => events.push(...e));
+                await partition.applyEvents([
+                    unlockSpace('admin', '3342', 123),
+                ]);
+
+                await waitAsync();
+
+                expect(events).toEqual([asyncResult(123, undefined)]);
+            });
+
+            it('should reject the async task if given the wrong password', async () => {
+                setupPartition({
+                    type: 'remote_causal_repo',
+                    branch: 'testBranch',
+                    host: 'testHost',
+                    static: true,
+                    readOnly: true,
+                });
+
+                const bot1 = atom(atomId('a', 1), null, bot('bot1'));
+                const tag1 = atom(atomId('a', 2), bot1, tag('tag1'));
+                const value1 = atom(atomId('a', 3), tag1, value('abc'));
+
+                partition.connect();
+
+                addAtoms.next({
+                    branch: 'testBranch',
+                    atoms: [bot1, tag1, value1],
+                });
+
+                let events = [] as Action[];
+                partition.onEvents.subscribe(e => events.push(...e));
+
+                await partition.applyEvents([
+                    unlockSpace('admin', 'wrong', 123),
+                ]);
+
+                await waitAsync();
+
+                expect(events).toEqual([
+                    asyncError(
+                        123,
+                        new Error(
+                            'Unable to unlock the space because the passcode is incorrect.'
+                        )
+                    ),
+                ]);
             });
         });